Tweak duscussion page format
